+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 =======
 ## [Unreleased]
 - Removed gonum.org/v1/gonum dependency.
+- Enable rpc-caller-procedure via yarpc outbound middleware.
 
 ## [1.71.0] - 2023-12-14
 - tchannel: optional transport-level config to allow reusing a buffer for reading a tchannel response body.
@@ -17,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.70.4] - 2023-08-31
 - logging: fix logged error in observability middleware when fields of transport.Request is in the tagsBlocklist
 - `go.mod`: update minimum requirements to go1.21 instead of go1.14 and update `golang.org/x/net v0.7.0` to v0.14.0
-- middleware stack usage: remove ~2KB of stack usage from the rpc handler function, 
+- middleware stack usage: remove ~2KB of stack usage from the rpc handler function,
   so that it decreases the chance of needing more stack. It can improve the
   performance of the application.